Refactoring utilities files names
